Allow generators to include custom trial metadata
Each metadata parameter will be added as trial attibute and
will be included in the history file.

The use case is that some generators (like the multitask generator)
need to attach custom fields to the libe_opt trials in order to relate
them to their own internal trials (e.g., Ax experiments have their own
way of organizing trials, which are not the same trials as those
managed by libe_opt). By including the metadata in the history file,
we allow also for the possibility of recreating the trials from this
file.