Add new `libe_comms` parameter.
Before, optimas attempted to detect whether the run had been launched
with an `mpirun` command or similar.
This required importing MPI, which then caused issued with openMPI in
local mode.
There seems to be no straightforward way of detecting if optimas was
launched with `mpirun`. The user will now have to specify if `local`
(default) or `mpi` communications will be used with the `libe_comms`
parameter.