Improve detection of parameter `value_type`.
The previous implementation was failing for numpy dtypes, as it was assuming the use of native python types. This prevented the generator
from being initalized, for example, from an `ExplorationDiagnostics`. This is because the parameters in the diags are read from the serialized json file, and thus always initialized with the numpy dtype name.